// du_website/src/Models/SalesInvoice.ts
// cat: "P"
// ​​​​
// currency: "LBP"
// ​​​​
// date_added: "2023-07-20T00:00:00.000Z"
// ​​​​
// invoice_no: "4448269"
// ​​​​
// is_paid: "partial"
// ​​​​
// oracle_number: "4243041"
// ​​​​
// remaining_amount: "529"
// ​​​​
// total_amount: "14845055"

type SalesInvoiceModel = {
  cat: string;
  currency: string;
  date_added: string;
  invoice_no: string;
  is_paid: string;
  oracle_number: string;
  remaining_amount: string;
  total_amount: string;
};
export default SalesInvoiceModel;